// src/services/types/platform/session_replay.ts
export type SessionReplayApiResponse= {
  _id: string
  __v: number
  session_replay: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
    }
    section3: {
      video1: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section6: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}